Fix MALCOLMLENNOX bug
The app wasn't accounting for a speech block that had two speakers speaking the same line, thus we we're getting a MALCOLMLENNOX speaker. This commit fixes that bug by iterating through each speech and pulling out the speakers, then putting each speaker into a hash with their lines as the values.